file refactoring: moved file into different directories in for managing completed tests
