+++ 1.Website/backend/src/test/Task.deleteTaskByTasknr.test.js
@@ -1,10 +1,14 @@
-const makeApp = require('../../app');
+const makeApp = require('../app');
 const supertest = require('supertest');
 const {MongoClient} = require('mongodb');
 
+/**** here we mock the database with the relevant collections and documents ***/
 
-describe('/insertTask',  ()=> {
-    describe('When request with a given JSON body ',  () =>{
+/**** here we mock the database with the relevant collections and documents  documents ***/
+
+describe('/deleteTaskByTasknr',  ()=> {
+
+    describe('When requested with a Project number' , ()=>{
         let connection;
         let MockDB;
         beforeAll(async () => {
@@ -33,47 +37,9 @@ describe('/insertTask',  ()=> {
             await connection.close();
             await MockDB.close();
         });
+
         it('it should return status code 200', async ()=> {
 
-            let MockTask = {
-                description: "Help Mark with his work",
-                status: "in-progress",
-                project: "Graph-Path",
-                tasknr: 1,
-                assignee: "Joe",
-                assigner: "Alistair",
-                due: Date.now(),
-                issued: Date.now()+48,
-            }
-            let app = makeApp(false,MockDB)
-            let response  = await supertest(app)
-                .post('/task/insertTask')
-                .send(MockTask)
-                .expect(200)
-                .then((res)=>{})
-        });
-        it('it should return a JSON object', async ()=> {
-
-            let MockTask = {
-                description: "Help Mark with his work",
-                status: "in-progress",
-                project: "Graph-Path",
-                tasknr: 1,
-                assignee: "Joe",
-                assigner: "Alistair",
-                due: Date.now(),
-                issued: Date.now()+48,
-            }
-            let app = makeApp(false,MockDB)
-            let response  = await supertest(app)
-                .post('/task/insertTask')
-                .send(MockTask)
-                .expect(200)
-                .then((res)=>{
-                    expect(res.headers['content-type']).toBe('application/json; charset=utf-8')
-                })
-        });
-        it('it should return JSON object with save message', async ()=> {
             const Tasks = MockDB.collection('Tasks');
             let MockTask = {
                 description: "Help Mark with his work",
@@ -85,22 +51,70 @@ describe('/insertTask',  ()=> {
                 due: Date.now(),
                 issued: Date.now()+48,
             }
-            await Tasks.insertOne(MockTask);
+
+            await Tasks.insertOne(MockTask)
             let app = makeApp(false,MockDB)
-            const response = await supertest(app)
-                .post('/task/insertTask')
-                .send(MockTask)
+            let response  = await supertest(app)
+                .delete('/task/deleteTaskByTasknr')
+                .send({
+                    tasknr: 1
+                })
+                .expect(200)
+                .then((res)=>{})
+        });
+        it('it should return a json body', async ()=> {
+
+            const Tasks = MockDB.collection('Tasks');
+            let MockTask = {
+                description: "Help Mark with his work",
+                status: "in-progress",
+                project: "Graph-Path",
+                tasknr: 1,
+                assignee: "Joe",
+                assigner: "Alistair",
+                due: Date.now(),
+                issued: Date.now()+48,
+            }
+
+            await Tasks.insertOne(MockTask)
+            let app = makeApp(false,MockDB)
+            let response  = await supertest(app)
+                .delete('/task/deleteTaskByTasknr')
+                .send({
+                    tasknr: 1
+                })
                 .expect(200)
                 .then((res)=>{
-                    res.body
-                    expect(res.body['message']).toBe('saved')
-
+                    expect(res.headers['content-type']).toBe('application/json; charset=utf-8')
                 })
         });
-    });
+        it('it should return message \'Deleted on success\'', async ()=> {
 
+            const Tasks = MockDB.collection('Tasks');
+            let MockTask = {
+                description: "Help Mark with his work",
+                status: "in-progress",
+                project: "Graph-Path",
+                tasknr: 1,
+                assignee: "Joe",
+                assigner: "Alistair",
+                due: Date.now(),
+                issued: Date.now()+48,
+            }
 
+            await Tasks.insertOne(MockTask)
+            let app = makeApp(false,MockDB)
+            let response  = await supertest(app)
+                .delete('/task/deleteTaskByTasknr')
+                .send({
+                    tasknr: 1
+                })
+                .expect(200)
+                .then((res)=>{
+                    expect(res.body['message']).toBe('Deleted')
+                })
+        });
 
-
+    })
 
 });
